fix(router): isolate the ssg worker bundle graph
