+++ packages/qwik-router/src/middleware/cloudflare-pages/index.ts
@@ -93,9 +93,9 @@ export function createQwikRouter(opts: QwikRouterCloudflarePagesOptions) {
       // send request to qwik router request handler
       const handledResponse = await requestHandler(serverRequestEv, opts);
       if (handledResponse) {
-        handledResponse.completion.then((v) => {
-          if (v) {
-            console.error(v);
+        handledResponse.completion.then((completion) => {
+          if (completion) {
+            console.error(completion);
           }
         });
         const response = await handledResponse.response;
