+++ packages/qwik-router/src/middleware/netlify-edge/index.ts
@@ -61,9 +61,9 @@ export function createQwikRouter(opts: QwikRouterNetlifyOptions) {
       // send request to qwik router request handler
       const handledResponse = await requestHandler(serverRequestEv, opts);
       if (handledResponse) {
-        handledResponse.completion.then((v) => {
-          if (v) {
-            console.error(v);
+        handledResponse.completion.then((completion) => {
+          if (completion) {
+            console.error(completion);
           }
         });
         const response = await handledResponse.response;
