+++ packages/qwik-router/src/middleware/request-handler/index.ts
@@ -10,7 +10,7 @@ export { ServerError } from './server-error';
 export { AbortMessage, RedirectMessage } from './redirect-handler';
 export { RewriteMessage } from './rewrite-handler';
 
-export { RequestEvShareQData } from './request-event';
+export { RequestEvShareQData } from './request-event-core';
 export { clearSsrCache } from './etag';
 export { _TextEncoderStream_polyfill } from './polyfill';
