+++ packages/qwik-router/src/middleware/vercel-edge/index.ts
@@ -88,9 +88,9 @@ export function createQwikRouter(opts: QwikRouterVercelEdgeOptions) {
       // send request to qwik router request handler
       const handledResponse = await requestHandler(serverRequestEv, opts);
       if (handledResponse) {
-        handledResponse.completion.then((v) => {
-          if (v) {
-            console.error(v);
+        handledResponse.completion.then((completion) => {
+          if (completion) {
+            console.error(completion);
           }
         });
         const response = await handledResponse.response;
