+++ packages/qwik-router/src/ssg/response-page-ssg.ts
@@ -1,4 +1,3 @@
-import { Q_ROUTE } from '../../runtime/src/constants';
 import {
   getRequestLoaders,
   getRequestLoaderSerializationStrategyMap,
@@ -8,8 +7,9 @@ import {
   RequestEvSharedActionId,
   RequestEvSharedNonce,
   RequestRouteName,
-} from './request-event-core';
-import { getQwikRouterServerDataWithDeps } from './response-page-core';
+} from '@qwik-router-ssg-worker/middleware/request-handler/request-event-core';
+import { getQwikRouterServerDataWithDeps } from '@qwik-router-ssg-worker/middleware/request-handler/response-page-core';
+import { Q_ROUTE } from './worker-imports/runtime';
 
 type GetQwikRouterServerDataArgs =
   Parameters<typeof getQwikRouterServerDataWithDeps> extends [any, ...infer Rest] ? Rest : never;
