+++ packages/qwik-router/src/ssg/worker-imports/runtime.ts
@@ -1,0 +1,10 @@
+export {
+  QACTION_KEY,
+  QDATA_KEY,
+  QFN_KEY,
+  QLOADER_KEY,
+  Q_ROUTE,
+} from '@qwik-router-ssg-worker/runtime/src/constants';
+export { resolveRouteConfig } from '@qwik-router-ssg-worker/runtime/src/head';
+export { loadRoute } from '@qwik-router-ssg-worker/runtime/src/routing';
+export { isPromise } from '@qwik-router-ssg-worker/runtime/src/utils';
