+++ tsconfig.json
@@ -18,6 +18,7 @@
     "noUnusedLocals": true,
     "paths": {
       "@qwik-client-manifest": ["packages/qwik/dist/server-modules.d.ts"],
+      "@qwik-router-ssg-worker/*": ["packages/qwik-router/src/*"],
       "@qwik-router-config": ["packages/qwik-router/src/runtime/src/qwik-router-config.ts"],
       "@qwik-router-sw-register-build": [
         "packages/qwik-router/src/buildtime/runtime-generation/sw-register-build.ts"
